Add files via upload
Through the GeoGKQuiz, people in areas that have strict Covid-19 regulations will still be able to educate themselves about the world they live in. With online teaching methods being implemented in recent time, this quiz will also be an effective tool for students learning
geography to use to study. Since the onset of the COVID-19 pandemic, children have been put
at greater risk of dropping out of school, lagging behind and losing learning (Sheperd, 2021).
With the implementation of our quiz, students will be given a more fun way of interacting with their work. For all of the reasons above the members of this group have decided that geography is a suitable topic for our quiz.
The quiz will consist of questions relating to capital cities, flags, national anthems, oceans, rivers, mountain ranges and fascinating monuments which are tourist hotspots. The GeoGKQuiz has the capacity to inform, educate and entertain which we believe is the best way to teach people and bring the world to them.
The quiz is console-based with time being recorded and will include a leader board for players to motivate themselves to do better.